+++ packages/core/src/components/MinimalButton.tsx
@@ -10,19 +10,19 @@ import * as React from 'react';
 
 import classNames from '../utils/classNames';
 
-interface ButtonProps {
+interface MinimalButtonProps {
     isDisabled?: boolean;
     isSelected?: boolean;
     onClick(): void;
 }
 
-const Button: React.FC<ButtonProps> = ({ children, isDisabled = false, isSelected = false, onClick }) => (
+const MinimalButton: React.FC<MinimalButtonProps> = ({ children, isDisabled = false, isSelected = false, onClick }) => (
     <button
         className={
             classNames({
-                'rpv-core__button': true,
-                'rpv-core__button--disabled': isDisabled,
-                'rpv-core__button--selected': isSelected,
+                'rpv-core__minimal-button': true,
+                'rpv-core__minimal-button--disabled': isDisabled,
+                'rpv-core__minimal-button--selected': isSelected,
             })
         }
         onClick={onClick}
@@ -31,4 +31,4 @@ const Button: React.FC<ButtonProps> = ({ children, isDisabled = false, isSelecte
     </button>
 );
 
-export default Button;
+export default MinimalButton;