+++ packages/core/src/index.ts
@@ -7,10 +7,10 @@
  */
 
 export { default as AnnotationType } from './annotations/AnnotationType';
-export { default as Button } from './components/Button';
 export { default as Menu } from './components/Menu';
 export { default as MenuDivider } from './components/MenuDivider';
 export { default as MenuItem } from './components/MenuItem';
+export { default as MinimalButton } from './components/MinimalButton';
 export { default as PrimaryButton } from './components/PrimaryButton';
 export { default as ProgressBar } from './components/ProgressBar';
 export { default as Separator } from './components/Separator';