+++ packages/page-navigation/src/GoToFirstPageButton.tsx
@@ -7,7 +7,7 @@
  */
 
 import * as React from 'react';
-import { Button, LocalizationContext, Position, Tooltip } from '@react-pdf-viewer/core';
+import { LocalizationContext, MinimalButton, Position, Tooltip } from '@react-pdf-viewer/core';
 
 import { RenderGoToPageProps } from './types/index';
 import UpArrowIcon from './UpArrowIcon';
@@ -22,7 +22,7 @@ const GoToFirstPageButton: React.FC<RenderGoToPageProps> = ({ isDisabled, onClic
     return (
         <Tooltip
             position={Position.BottomCenter}
-            target={<Button isDisabled={isDisabled} onClick={onClick}><UpArrowIcon /></Button>}
+            target={<MinimalButton isDisabled={isDisabled} onClick={onClick}><UpArrowIcon /></MinimalButton>}
             content={() => label}
             offset={TOOLTIP_OFFSET}
         />